1.更新

+++ basic/configSoft.sh
@@ -8,11 +8,18 @@ rm -fr ../$softName
 mv -f * ../
 cd ../
 rm -fr $unzipPath
-#copy default config file
-rm -fr $installPath/$softName/etc/*
-cp -r /root/git/shell/$1/config/default/* $installPath/$softName/etc/
-#special config
-for shell in `find /root/git/shell/$1/config/special/*`
-	{ 
-		bash $shell || exit $?
-	}
+if [ $1 = "zk" -o $1 = "ZK" ]
+	then
+		rm -fr $installPath/$softName/conf/*
+		cp -r /root/git/shell/$1/config/default/* $installPath/$softName/conf/ || { echo "cp ZK default config error";exit -18; }
+	else
+		#copy default config file
+		rm -fr $installPath/$softName/etc/*
+		cp -r /root/git/shell/$1/config/default/* $installPath/$softName/etc/ || { echo "cp default config error";exit -19; }
+		#special config
+		find /root/git/shell/$1/config/special/* || exit 0
+		for shell in `find /root/git/shell/$1/config/special/*`
+			{ 
+				bash $shell || exit $?
+			}
+fi
